Renamed app subgraph template and updated it and voting subgraphs with latest architecture from the token manager subgraph
