--- packages/app-subgraph/src/config.ts
@@ -1,2 +1,0 @@
-export const APP_ID = '0x9fa3927f639745e587912d4b0fea7ef9013bf93fb907d29faeab57417ba6e1d4'
-export const APP_TEMPLATE_NAME = 'Voting'
