+++ packages/app-template-subgraph/abis/Voting.json
