+++ packages/app-template-subgraph/manifest/data/mainnet-staging.json
