+++ packages/app-template-subgraph/manifest/data/mainnet.json
