+++ packages/app-template-subgraph/manifest/data/rinkeby-staging.json
