+++ packages/app-template-subgraph/manifest/data/rinkeby.json
