+++ packages/app-template-subgraph/manifest/templates/Kernel.template.yaml
@@ -19,11 +19,11 @@
     apiVersion: 0.0.1
     language: wasm/assemblyscript
     entities:
-      - Organization
+      - NoEnt
     abis:
       - name: Kernel
         file: ./abis/Kernel.json
     eventHandlers:
       - event: NewAppProxy(address,bool,bytes32)
         handler: handleNewAppProxy
-    file: ./src/mappings/base/AragonApp.ts
+    file: ./src/base/DAOFactory.ts
