+++ packages/app-template-subgraph/schema.graphql
