+++ packages/app-template-subgraph/scripts/build-manifest.sh
