+++ packages/app-template-subgraph/scripts/deploy.sh
