+++ packages/token-manager-subgraph/manifest/templates/Kernel.template.yaml
@@ -19,7 +19,7 @@
     apiVersion: 0.0.1
     language: wasm/assemblyscript
     entities:
-      - Organization
+      - NoEnt
     abis:
       - name: Kernel
         file: ./abis/Kernel.json
