+++ packages/voting-subgraph/manifest/templates/DAOFactory.template.yaml
@@ -10,11 +10,11 @@
     apiVersion: 0.0.1
     language: wasm/assemblyscript
     entities:
-      - Organization
+      - NoEnt
     abis:
       - name: DAOFactory
         file: ./abis/DAOFactory.json
     eventHandlers:
       - event: DeployDAO(address)
         handler: handleDeployDAO
-    file: ./src/mappings/base/AragonOrg.ts
+    file: ./src/base/DAOFactory.ts
